Ignore log files from the current month
Only log files from previous months should be archived

+++ archive.sh
@@ -3,7 +3,7 @@
 # (C) Alex Vinall 2017
 # Archives log files into a 7-zip archive per month, using Ultra compression.
 # Intended to be used with log files where the filename ends with:
-# ${current_date}.out
+# YYYYMMDD_HHMM.out
 
 # Archives log files into 7z archives per month
 echo "Archiving log files..."
@@ -22,15 +22,24 @@ else
   do
     year=$(echo $i | cut -c1-4)
     month=$(echo $i | cut -c5-6)
-    echo "Archiving $year-$month..."
-    status=$(7z a -t7z -mx=9 Archive/$year-$month-Logs.7z *$i*.out | tail -n 1)
-    echo "7-zip reports: $status"
-    if [ "$status" == "Everything is Ok" ]
+
+    # Ignore any files from the current month
+    if [ `date '+%Y-%m'` = $year-$month ]
     then
-      echo "$year-$month: archived OK, removing log files."
-      rm *$i*.out
-    else
-      echo "$year-$month: archive error, log files will not be removed."
+      echo "$year-$month: ignoring log files from the current month."
+    
+    # Archive all previous months in a 7z per month
+    else  
+      echo "Archiving $year-$month..."
+      status=$(7z a -t7z -mx=9 Archive/$year-$month-Logs.7z *$i*.out | tail -n 1)
+      echo "7-zip reports: $status"
+      if [ "$status" == "Everything is Ok" ]
+      then
+        echo "$year-$month: archived OK, removing log files."
+        rm *$i*.out
+      else
+        echo "$year-$month: archive error, log files will not be removed."
+      fi
     fi
   done
 fi
